routes: migrate roles domain (9 routes)
Move to src/server/routes/roles.ts:
- /api/roles/assistant/prompts (GET, PUT :type)
- /api/roles (GET list, POST create)
- /api/roles/:name (GET, PUT, DELETE)
- /api/roles/:name/customize POST
- /api/roles/:name/override DELETE

Net: server.ts -258 lines (7053 → 6795). No behaviour change.
