+++ src/server/routes/dispatcher.ts
@@ -21,6 +21,7 @@ import { oauthRoutes } from "./oauth.js";
 import { imageGenerationRoutes } from "./image-generation.js";
 import { modelsRoutes } from "./models.js";
 import { preferencesConfigRoutes } from "./preferences-config.js";
+import { rolesRoutes } from "./roles.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -33,6 +34,7 @@ const allRoutes: ReadonlyArray<Route> = [
 	...imageGenerationRoutes,
 	...modelsRoutes,
 	...preferencesConfigRoutes,
+	...rolesRoutes,
 ];
 
 /**
