+++ src/server/server.ts
@@ -1076,7 +1076,7 @@ async function handleApiRoute(
 ) {
 	// These are always wired by the sole caller; the optional markers are only to avoid
 	// touching every existing signature site.
-	const serverRoleStore = roleStore!;
+	void roleStore;
 	const json = (data: unknown, status = 200) => {
 		res.writeHead(status, { "Content-Type": "application/json" });
 		res.end(JSON.stringify(data));
@@ -2949,264 +2949,6 @@ async function handleApiRoute(
 	// ── Config: default cwd ──
 
 	// ── Unified Model Registry ──
-
-	// GET /api/roles/assistant/prompts — must come before :name route
-	if (url.pathname === "/api/roles/assistant/prompts" && req.method === "GET") {
-		const { ASSISTANT_REGISTRY } = await import("./agent/assistant-registry.js");
-		const prompts = Object.values(ASSISTANT_REGISTRY).map((def) => ({
-			type: def.type,
-			title: def.title,
-			promptTitle: def.promptTitle,
-			prompt: def.prompt,
-		}));
-		json({ prompts });
-		return;
-	}
-
-	// PUT /api/roles/assistant/prompts/:type
-	if (url.pathname.startsWith("/api/roles/assistant/prompts/") && req.method === "PUT") {
-		const type = url.pathname.slice("/api/roles/assistant/prompts/".length);
-		if (!type) {
-			json({ error: "Missing type parameter" }, 400);
-			return;
-		}
-		const body = await readBody(req);
-		const { updateAssistantDef } = await import("./agent/assistant-registry.js");
-		const updated = updateAssistantDef(type, {
-			prompt: body?.prompt,
-			title: body?.title,
-			promptTitle: body?.promptTitle,
-		});
-		if (!updated) {
-			json({ error: `Unknown assistant type: ${type}` }, 404);
-			return;
-		}
-		json(updated);
-		return;
-	}
-
-	// GET /api/roles (with cascade origin)
-	if (url.pathname === "/api/roles" && req.method === "GET") {
-		const projectId = url.searchParams.get("projectId") || undefined;
-		const resolved = configCascade.resolveRoles(projectId);
-		json({ roles: resolved.map(r => ({ ...r.item, origin: r.origin, ...(r.overrides ? { overrides: r.overrides } : {}) })) });
-		return;
-	}
-
-	// POST /api/roles (scope-aware: body.projectId → create in that project's store)
-	if (url.pathname === "/api/roles" && req.method === "POST") {
-		const body = await readBody(req);
-		const targetProjectId = body?.projectId;
-		try {
-			if (targetProjectId) {
-				const ctx = projectContextManager.getOrCreate(targetProjectId);
-				if (!ctx) { json({ error: "Project not found" }, 404); return; }
-				const now = Date.now();
-				const role = {
-					name: body?.name,
-					label: body?.label ?? body?.name,
-					promptTemplate: body?.promptTemplate || "",
-					accessory: body?.accessory ?? "none",
-					toolPolicies: body?.toolPolicies,
-					model: typeof body?.model === "string" && body.model.trim() ? body.model.trim() : undefined,
-					thinkingLevel: typeof body?.thinkingLevel === "string" && body.thinkingLevel.trim() ? body.thinkingLevel.trim() : undefined,
-					createdAt: now,
-					updatedAt: now,
-				};
-				if (!role.name || typeof role.name !== "string") throw new Error("Missing name");
-				const NAME_PATTERN = /^[a-z0-9][a-z0-9-]*[a-z0-9]$|^[a-z0-9]$/;
-				if (!NAME_PATTERN.test(role.name)) throw new Error("Role name must be lowercase alphanumeric + hyphens");
-				ctx.roleStore.put(role);
-				json(role, 201);
-			} else {
-				const role = roleManager.createRole({
-					name: body?.name,
-					label: body?.label,
-					promptTemplate: body?.promptTemplate || "",
-					accessory: body?.accessory,
-					toolPolicies: body?.toolPolicies,
-					model: typeof body?.model === "string" && body.model.trim() ? body.model.trim() : undefined,
-					thinkingLevel: typeof body?.thinkingLevel === "string" && body.thinkingLevel.trim() ? body.thinkingLevel.trim() : undefined,
-				});
-				json(role, 201);
-			}
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// POST /api/roles/:name/customize — copy resolved role to a target scope
-	const roleCustomizeMatch = url.pathname.match(/^\/api\/roles\/([^/]+)\/customize$/);
-	if (roleCustomizeMatch && req.method === "POST") {
-		const name = decodeURIComponent(roleCustomizeMatch[1]);
-		const scope = url.searchParams.get("scope") || "server";
-		const projectId = url.searchParams.get("projectId") || undefined;
-
-		const resolved = configCascade.resolveRoles(projectId);
-		const source = resolved.find(r => r.item.name === name);
-		if (!source) { json({ error: "Role not found" }, 404); return; }
-
-		let targetStore;
-		if (scope === "project") {
-			if (!projectId) { json({ error: "projectId required for project scope" }, 400); return; }
-			const ctx = projectContextManager.getOrCreate(projectId);
-			if (!ctx) { json({ error: "Project not found" }, 404); return; }
-			targetStore = ctx.roleStore;
-		} else {
-			// scope === "server" (or unspecified) → system/server layer
-			targetStore = serverRoleStore;
-		}
-
-		const now = Date.now();
-		const copy = { ...source.item, createdAt: now, updatedAt: now };
-		targetStore.put(copy);
-		json(copy, 201);
-		return;
-	}
-
-	// DELETE /api/roles/:name/override — remove override at a scope, reverting to inherited
-	const roleOverrideMatch = url.pathname.match(/^\/api\/roles\/([^/]+)\/override$/);
-	if (roleOverrideMatch && req.method === "DELETE") {
-		const name = decodeURIComponent(roleOverrideMatch[1]);
-		const scope = url.searchParams.get("scope") || "server";
-		const projectId = url.searchParams.get("projectId") || undefined;
-
-		let targetStore;
-		if (scope === "project") {
-			if (!projectId) { json({ error: "projectId required for project scope" }, 400); return; }
-			const ctx = projectContextManager.getOrCreate(projectId);
-			if (!ctx) { json({ error: "Project not found" }, 404); return; }
-			targetStore = ctx.roleStore;
-		} else {
-			// scope === "server" (or unspecified) → system/server layer
-			targetStore = serverRoleStore;
-		}
-
-		targetStore.remove(name);
-		json({ ok: true });
-		return;
-	}
-
-	// Routes with role :name parameter
-	const roleMatch = url.pathname.match(/^\/api\/roles\/([^/]+)$/);
-	if (roleMatch) {
-		const name = decodeURIComponent(roleMatch[1]);
-
-		if (req.method === "GET") {
-			const qProjectId = url.searchParams.get("projectId") || undefined;
-			if (qProjectId) {
-				const resolved = configCascade.resolveRoles(qProjectId);
-				const found = resolved.find(r => r.item.name === name);
-				if (!found) { json({ error: "Role not found" }, 404); return; }
-				json({ ...found.item, origin: found.origin, ...(found.overrides ? { overrides: found.overrides } : {}) });
-			} else {
-				const role = roleManager.getRole(name);
-				if (!role) { json({ error: "Role not found" }, 404); return; }
-				json(role);
-			}
-			return;
-		}
-
-		if (req.method === "PUT") {
-			const body = await readBody(req);
-			if (!body) { json({ error: "Missing body" }, 400); return; }
-			const qProjectId = url.searchParams.get("projectId") || undefined;
-			if (qProjectId) {
-				const ctx = projectContextManager.getOrCreate(qProjectId);
-				if (!ctx) { json({ error: "Project not found" }, 404); return; }
-				const existing = ctx.roleStore.get(name);
-				if (!existing) { json({ error: "Role not found in project" }, 404); return; }
-				const validPolicies = new Set(['allow', 'ask', 'never', 'always-allow', 'ask-once', 'always-ask', 'never-ask']);
-				let toolPolicies = existing.toolPolicies;
-				if (body.toolPolicies !== undefined) {
-					const cleaned: Record<string, any> = {};
-					if (body.toolPolicies && typeof body.toolPolicies === 'object') {
-						for (const [k, v] of Object.entries(body.toolPolicies)) {
-							if (typeof v === 'string' && validPolicies.has(v)) cleaned[k] = v;
-						}
-					}
-					toolPolicies = cleaned;
-				}
-				// model / thinkingLevel: explicit empty string clears the field; absent leaves unchanged.
-				let model = existing.model;
-				if (body.model !== undefined) {
-					model = typeof body.model === "string" && body.model.trim() ? body.model.trim() : undefined;
-				}
-				let thinkingLevel = existing.thinkingLevel;
-				if (body.thinkingLevel !== undefined) {
-					thinkingLevel = typeof body.thinkingLevel === "string" && body.thinkingLevel.trim() ? body.thinkingLevel.trim() : undefined;
-				}
-				const updated = {
-					...existing,
-					label: body.label ?? existing.label,
-					promptTemplate: body.promptTemplate ?? existing.promptTemplate,
-					accessory: body.accessory ?? existing.accessory,
-					toolPolicies,
-					model,
-					thinkingLevel,
-					name,
-					updatedAt: Date.now(),
-				};
-				ctx.roleStore.put(updated);
-				json({ ok: true });
-			} else {
-				// model / thinkingLevel: explicit empty string clears the field; absent leaves unchanged.
-				const modelUpdate = body.model !== undefined
-					? (typeof body.model === "string" && body.model.trim() ? body.model.trim() : "")
-					: undefined;
-				const thinkingUpdate = body.thinkingLevel !== undefined
-					? (typeof body.thinkingLevel === "string" && body.thinkingLevel.trim() ? body.thinkingLevel.trim() : "")
-					: undefined;
-				// Apply model/thinking via direct store update to support clearing (yaml-store update treats undefined as "don't change").
-				if (modelUpdate !== undefined || thinkingUpdate !== undefined) {
-					const existing = roleManager.getRole(name);
-					if (existing) {
-						const patched = {
-							...existing,
-							model: modelUpdate !== undefined ? (modelUpdate || undefined) : existing.model,
-							thinkingLevel: thinkingUpdate !== undefined ? (thinkingUpdate || undefined) : existing.thinkingLevel,
-							updatedAt: Date.now(),
-						};
-						serverRoleStore.put(patched);
-					}
-				}
-				const ok = roleManager.updateRole(name, {
-					label: body.label,
-					promptTemplate: body.promptTemplate,
-					accessory: body.accessory,
-					toolPolicies: body.toolPolicies !== undefined ? (() => {
-						const validPolicies = new Set(['allow', 'ask', 'never', 'always-allow', 'ask-once', 'always-ask', 'never-ask']);
-						const cleaned: Record<string, import("./agent/role-store.js").GrantPolicy> = {};
-						if (body.toolPolicies && typeof body.toolPolicies === 'object') {
-							for (const [k, v] of Object.entries(body.toolPolicies)) {
-								if (typeof v === 'string' && validPolicies.has(v)) cleaned[k] = v as import("./agent/role-store.js").GrantPolicy;
-							}
-						}
-						return cleaned;
-					})() : undefined,
-				});
-				if (!ok) { json({ error: "Role not found" }, 404); return; }
-				json({ ok: true });
-			}
-			return;
-		}
-
-		if (req.method === "DELETE") {
-			const qProjectId = url.searchParams.get("projectId") || undefined;
-			if (qProjectId) {
-				const ctx = projectContextManager.getOrCreate(qProjectId);
-				if (!ctx) { json({ error: "Project not found" }, 404); return; }
-				ctx.roleStore.remove(name);
-				json({ ok: true });
-			} else {
-				const ok = roleManager.deleteRole(name);
-				if (!ok) { json({ error: "Role not found" }, 404); return; }
-				json({ ok: true });
-			}
-			return;
-		}
-	}
 
 	// ── Task endpoints ─────────────────────────────────────────────
